// libs/settings-dialog/src/lib/model/settings.ts
/*
 *  Copyright (c) 2020 Robert Bosch Manufacturing Solutions GmbH, Germany. All rights reserved.
 */
export interface Settings {
  showEditorMap: boolean;
  showEditorNav: boolean;
  autoSaveEnabled: boolean;
  autoValidationEnabled: boolean;
  showValidationPopupNotifications: boolean;
  enableHierarchicalLayout: boolean;
  validationTimerSeconds: number;
  saveTimerSeconds: number;
  showEntityValueEntityEdge: boolean;
  showConnectionLabels: boolean;
}